+++ src/components/Thumbnail/Thumbnail.stories.tsx
@@ -1,13 +1,15 @@
 import { action } from '@storybook/addon-actions';
 import { boolean, text } from '@storybook/addon-knobs';
 import React from 'react';
+import close from '../../icons/close-24px.svg';
+import rotate from '../../icons/rotate_right-24px.svg';
 import { createFile, CreateFileOptions } from '../../storybook-helpers/data/files';
 import Thumbnail, { ThumbnailProps } from '../Thumbnail';
 import docs from './README.md';
 
 export default { title: 'Thumbnail', component: Thumbnail, parameters: { info: docs } };
 
-const defaultProps = (options?: CreateFileOptions, index = 0): ThumbnailProps => ({
+const defaultProps = (options?: CreateFileOptions, index = 0, withToolButtons?: boolean): ThumbnailProps => ({
   file: createFile(index, options),
   selected: boolean('selected', false),
   hideExtension: boolean('hideExtension', false),
@@ -16,8 +18,12 @@ const defaultProps = (options?: CreateFileOptions, index = 0): ThumbnailProps =>
   otherDragging: boolean('otherDragging', false),
   onClick: boolean('has onClick', true) ? action('onClick') : undefined,
   onRename: boolean('has onRename', true) ? action('onRename') : undefined,
-  onRemove: boolean('has onRemove', true) ? action('onRemove') : undefined,
-  onRotate: boolean('has onRotate', true) ? action('onRotate') : undefined,
+  buttonProps: withToolButtons
+    ? [
+        { children: <img src={rotate} alt={'rotate'} />, onClick: action('rotate onClick') },
+        { children: <img src={close} alt={'close'} />, onClick: action('close onClick') },
+      ]
+    : undefined,
 });
 
 export const basic = () => <Thumbnail {...defaultProps()} />;
@@ -25,6 +31,8 @@ export const basic = () => <Thumbnail {...defaultProps()} />;
 export const throttled = () => <Thumbnail {...defaultProps({ lazy: true })} />;
 
 export const pending = () => <Thumbnail {...defaultProps({ pending: true })} />;
+
+export const withToolButtons = () => <Thumbnail {...defaultProps(undefined, undefined, true)} />;
 
 export const withLabel = () => <Thumbnail label={text('label', 'some_label')} {...defaultProps()} />;